% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_site.R
\name{dataset_template_BGS_site}
\alias{dataset_template_BGS_site}
\title{create dataset template from BGS site file}
\usage{
dataset_template_BGS_site(files)
}
\arguments{
\item{files}{list with data and fasta paths}
}
\value{
data.frame with annotation template
}
\description{
create dataset template from BGS site file
}
\examples{
# Use package test data
path <- system.file("extdata", "BGS_site", package = "prolfquappPTMreaders")
files <- get_BGS_site_files(path)
annot_template <- dataset_template_BGS_site(files)
head(annot_template)

}
