% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_combined_STY.R
\name{dataset_template_FP_combined_STY}
\alias{dataset_template_FP_combined_STY}
\title{create dataset template for FP combined STY file}
\usage{
dataset_template_FP_combined_STY(files)
}
\arguments{
\item{files}{list with data and fasta paths}
}
\value{
data.frame with annotation template
}
\description{
create dataset template for FP combined STY file
}
