% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_combined_STY.R
\name{dataset_template_FP_combined_STY_v2}
\alias{dataset_template_FP_combined_STY_v2}
\title{create dataset template for FP combined STY file (v2 with manifest)}
\usage{
dataset_template_FP_combined_STY_v2(files)
}
\arguments{
\item{files}{list with data, fasta and fp.manifest paths}
}
\value{
data.frame with annotation template
}
\description{
create dataset template for FP combined STY file (v2 with manifest)
}
