% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_multisite.R
\name{dataset_template_FP_multi_site}
\alias{dataset_template_FP_multi_site}
\title{create dataset template from FP multi_site}
\usage{
dataset_template_FP_multi_site(files)
}
\description{
create dataset template from FP multi_site
}
