% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSSTATS_site.R
\name{dataset_template_MSSTATS_site}
\alias{dataset_template_MSSTATS_site}
\title{Annotation template for a site table}
\usage{
dataset_template_MSSTATS_site(files)
}
\arguments{
\item{files}{output of [get_MSSTATS_site_files()].}
}
\value{
data.frame with annotation template
}
\description{
Annotation template for a site table
}
\examples{
path <- system.file("extdata", "MSSTATS_site", package = "prolfquappPTMreaders")
dataset_template_MSSTATS_site(get_MSSTATS_site_files(path))
}
