% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_multisite.R
\name{get_FP_single_site_files}
\alias{get_FP_single_site_files}
\title{get report.tsv and fasta file location in folder}
\usage{
get_FP_single_site_files(path)
}
\value{
list with paths to data and fasta
}
\description{
get report.tsv and fasta file location in folder
}
\examples{
identical(formals(get_FP_multi_site_files), formals(prolfquapp::get_dummy_files))

}
