% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_site.R
\name{preprocess_BGS_site}
\alias{preprocess_BGS_site}
\title{preprocess Spectronaut PTM site data}
\usage{
preprocess_BGS_site(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_"
)
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess Spectronaut PTM site data
}
\examples{
identical(names(formals(preprocess_BGS_site)), names(formals(prolfquapp::preprocess_dummy)))
path <- "/Users/wolski/Dropbox/DataAnalysis/o40094_Fabienne/specnaut_outputs/SA001PT/20251104_093117_20250610-134014_SA001PO_Report_WPR"
files <- get_BGS_site_files(path)

# Create annotation
annot <- dataset_template_BGS_site(files)
# Fill in Group and Control columns
# Group A = Control (C), Group B = Treatment (T)
n_samples <- nrow(annot)
annot$Group <- rep(c("A", "B"), length.out = n_samples)
annot$Control <- ifelse(annot$Group == "A", "C", "T")
annotation <- prolfquapp::read_annotation(annot)

# Preprocess
result <- preprocess_BGS_site(
  quant_data = files$data,
  fasta_file = files$fasta,
  annotation = annotation
)
result$lfqdata
result$protein_annotation

}
