% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSSTATS_site.R
\name{preprocess_MSSTATS_site}
\alias{preprocess_MSSTATS_site}
\title{Preprocess a site-level table in MSstats long format}
\usage{
preprocess_MSSTATS_site(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_"
)
}
\arguments{
\item{quant_data}{path to the site table.}

\item{fasta_file}{path to the FASTA the windows are cut from.}

\item{annotation}{output of [prolfquapp::read_annotation()].}

\item{pattern_contaminants}{contaminant pattern.}

\item{pattern_decoys}{decoy pattern.}
}
\value{
list with lfqdata and protein annotation
}
\description{
Preprocess a site-level table in MSstats long format
}
\examples{
path <- system.file("extdata", "MSSTATS_site", package = "prolfquappPTMreaders")
files <- get_MSSTATS_site_files(path)
annot <- dataset_template_MSSTATS_site(files)
annot$Group <- rep(c("A", "B"), length.out = nrow(annot))
annot$Control <- ifelse(annot$Group == "A", "C", "T")
annotation <- prolfquapp::read_annotation(annot)
result <- preprocess_MSSTATS_site(
  quant_data = files$data,
  fasta_file = files$fasta,
  annotation = annotation
)
stopifnot(nrow(result$lfqdata$data) > 0)
head(result$protein_annotation$row_annot$SequenceWindow)
}
