% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolfqua_preprocess_functions.R
\docType{data}
\name{prolfqua_preprocess_functions}
\alias{prolfqua_preprocess_functions}
\title{preprocess functions implemented in this package}
\format{
An object of class \code{list} of length 5.
}
\usage{
prolfqua_preprocess_functions
}
\description{
preprocess functions implemented in this package
}
\keyword{datasets}
