% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_multisite.R
\name{read_FP_multisite_to_long}
\alias{read_FP_multisite_to_long}
\title{read multisite file}
\usage{
read_FP_multisite_to_long(quant_data)
}
\arguments{
\item{quant_data}{path to multisite quantification file}
}
\value{
tidy long-format data frame
}
\description{
read multisite file
}
