% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MSSTATS_site.R
\name{read_MSSTATS_site}
\alias{read_MSSTATS_site}
\title{Read the site table}
\usage{
read_MSSTATS_site(file)
}
\arguments{
\item{file}{path to the table; comma or tab separated.}
}
\value{
data frame with `PeptideSequence` filled in where absent.
}
\description{
Read the site table
}
\examples{
path <- system.file("extdata", "MSSTATS_site", package = "prolfquappPTMreaders")
sites <- read_MSSTATS_site(get_MSSTATS_site_files(path)$data)
head(sites)
}
