% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_site.R
\name{dataset_template_BGS_site}
\alias{dataset_template_BGS_site}
\title{create dataset template from BGS site file}
\usage{
dataset_template_BGS_site(files)
}
\arguments{
\item{files}{list with data and fasta paths}
}
\value{
data.frame with annotation template
}
\description{
create dataset template from BGS site file
}
\examples{
\dontrun{
path <- "/Users/wolski/Dropbox/DataAnalysis/o40094_Fabienne/specnaut_outputs/SA001PT/20251104_093117_20250610-134014_SA001PO_Report_WPR"
files <- get_BGS_site_files(path)
annot_template <- dataset_template_BGS_site(files)
head(annot_template)
}
}
