% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_combined_STY.R
\name{dataset_template_FP_combined_STY_v2}
\alias{dataset_template_FP_combined_STY_v2}
\title{create dataset tempalte for FP combined STY file}
\usage{
dataset_template_FP_combined_STY_v2(files)
}
\value{
data.frame
}
\description{
create dataset tempalte for FP combined STY file
}
