% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_multisite.R
\name{dataset_template_FP_multi_site}
\alias{dataset_template_FP_multi_site}
\title{create dataset template from FP multi_site}
\usage{
dataset_template_FP_multi_site(files)
}
\arguments{
\item{files}{list with data and fasta paths}
}
\value{
data.frame with annotation template
}
\description{
create dataset template from FP multi_site
}
\examples{
path <- system.file("extdata", "FP_multisite", package = "prolfquappPTMreaders")
files <- get_FP_multi_site_files(path)
annot_template <- dataset_template_FP_multi_site(files)
annot_template

}
