% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_site.R
\name{get_BGS_site_files}
\alias{get_BGS_site_files}
\title{get Spectronaut report.tsv and fasta file location in folder}
\usage{
get_BGS_site_files(path)
}
\arguments{
\item{path}{directory path to search for files}
}
\value{
list with paths to data and fasta
}
\description{
get Spectronaut report.tsv and fasta file location in folder
}
\examples{
identical(formals(get_BGS_site_files), formals(prolfquapp::get_dummy_files))
# Use package test data
path <- system.file("extdata", "BGS_site", package = "prolfquappPTMreaders")
files <- get_BGS_site_files(path)
files$data
files$fasta

}
