% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_site.R
\name{preprocess_BGS_site}
\alias{preprocess_BGS_site}
\title{preprocess Spectronaut PTM site data}
\usage{
preprocess_BGS_site(
  quant_data,
  fasta_file,
  annotation,
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_"
)
}
\arguments{
\item{quant_data}{path to Spectronaut PTM report file}

\item{fasta_file}{path to FASTA file}

\item{annotation}{annotation object from prolfquapp::read_annotation()}

\item{pattern_contaminants}{regex pattern to identify contaminant proteins}

\item{pattern_decoys}{regex pattern to identify decoy proteins}
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess Spectronaut PTM site data
}
