% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_FP_multisite.R
\name{preprocess_FP_multi_site}
\alias{preprocess_FP_multi_site}
\title{preprocess FP multisite, filter by purity_threshold and PeptideProphetProb}
\usage{
preprocess_FP_multi_site(
  quant_data,
  fasta_file,
  annotation,
  sitetype = c("singlesite", "multisite"),
  pattern_contaminants = "^zz|^CON|Cont_",
  pattern_decoys = "^REV_|^rev_"
)
}
\arguments{
\item{quant_data}{path to multisite quantification file}

\item{fasta_file}{path to FASTA file}

\item{annotation}{annotation object from prolfquapp::read_annotation()}

\item{sitetype}{type of site analysis ("singlesite" or "multisite")}

\item{pattern_contaminants}{regex pattern to identify contaminant proteins}

\item{pattern_decoys}{regex pattern to identify decoy proteins}
}
\value{
list with lfqdata and protein annotation
}
\description{
preprocess FP multisite, filter by purity_threshold and PeptideProphetProb
}
\examples{
identical(names(formals(preprocess_FP_multi_site)), names(formals(prolfquapp::preprocess_dummy)))

}
