% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_BGS_site.R
\name{read_BGS_site}
\alias{read_BGS_site}
\title{read Spectronaut PTM report file}
\usage{
read_BGS_site(quant_data, min_site_sample_loc = 0.3, min_site_loc = 0.95)
}
\arguments{
\item{quant_data}{path to Spectronaut PTM report file}

\item{min_site_sample_loc}{minimum site probability threshold for individual observations (default: 0.00)}

\item{min_site_loc}{minimum site probability threshold - keep only sites where at least one sample exceeds this (default: 0.99)}
}
\value{
data table with Spectronaut column names (dots replaced with underscores),
  filtered to keep only single phosphorylation sites (PTM_Multiplicity == 1) and high confidence observations
}
\description{
read Spectronaut PTM report file
}
\examples{
path <- "/Users/wolski/Dropbox/DataAnalysis/o40094_Fabienne/specnaut_outputs/SA001PT/20251104_093117_20250610-134014_SA001PO_Report_WPR"
files <- get_BGS_site_files(path)
data <- read_BGS_site(files$data, min_site_loc=0.99999)
nrow(data)
#View(data)
range(data$PTM_SiteProbability)
unique(data$PTM_ModificationTitle)
}
